requirements.txt
pytest==7.2.1
selene==2.0.0b16
allure-pytest==2.12.0
python-dotenv==0.21.0
Appium-Python-Client==2.6.1
pydantic==1.10.5